fix(FEC-14496): sync volume map with media (#46)
